Move pkg/configmap to pkg/config/ui
